+++ src/Modules/StoreUser/StoreUserControl.php
@@ -289,6 +289,7 @@ class StoreUserControl extends Control
 							'storeId' => $storeId,
 							'storeManagers' => $this->storeGateway->getStoreManagers($storeId),
 							'mayWritePost' => $this->storePermissions->mayWriteStoreWall($storeId),
+							'mayDeleteEverything' => $this->storePermissions->mayDeleteStoreWall($storeId),
 						])
 					);
 				} else {